Extract upgrade tests to their own module.
This fixes the test failure where the upgrade tests run the `pre_test_setup`
fixture, when they shouldn't.
